+++ macros/tests/gprogram_tests.rs
@@ -1,7 +1,7 @@
-mod program_basic;
+mod gprogram_basic;
 
 #[test]
 fn gprogram_basic() {
-    use program_basic::MyProgram;
+    use gprogram_basic::MyProgram;
     let _prg = MyProgram::new(42);
 }
